+++ cx-portal/src/services/AccessService.tsx
@@ -61,9 +61,10 @@ import { EnableIDPSuccess } from 'components/overlays/EnableIDP/EnableIDPSuccess
 import { DisableIDP } from 'components/overlays/EnableIDP/DisableIDP'
 import { AddusersIDP } from 'components/overlays/AddusersIDP'
 import AddServiceProvider from 'components/overlays/AddServiceProvider'
-import DeclineAdminboard from 'components/overlays/DeclineAdminboard'
 import EditPortalRoles from 'components/overlays/EditPortalRoles'
 import { UpdateIDPSuccess } from 'components/overlays/UpdateIDPSuccess'
+import ServiceDeclineAdminboard from 'components/overlays/DeclineAdminboard/ServiceDeclineAdminboard'
+import AppDeclineAdminboard from 'components/overlays/DeclineAdminboard/AppDeclineAdminboard'
 
 let pageMap: { [page: string]: IPage }
 let actionMap: { [action: string]: IAction }
@@ -186,8 +187,10 @@ export const getOverlay = (overlay: OverlayState) => {
       return <SampleForm />
     case OVERLAYS.ADD_SERVICE_PROVIDER:
       return <AddServiceProvider />
-    case OVERLAYS.DECLINE_ADMINBOARD:
-      return <DeclineAdminboard id={overlay.id} />
+    case OVERLAYS.SERVICE_DECLINE_ADMINBOARD:
+      return <ServiceDeclineAdminboard id={overlay.id} />
+    case OVERLAYS.APP_DECLINE_ADMINBOARD:
+      return <AppDeclineAdminboard id={overlay.id} />
     case OVERLAYS.EDIT_PORTAL_ROLES:
       return <EditPortalRoles id={overlay.id} />
     default: